Clean: hook fncs, add hook.c
